trying to fix the sign up
