+++ src/components/CoachSignUp.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { CoachSignUpp } from './coachsignup/CoachSignUpp';
+
+import { Header } from './shared/Header';
+
+
+export const CoachSignUp = () => {
+    return (
+        <>
+        <Header></Header>
+        <CoachSignUpp></CoachSignUpp>
+        </>
+    );
+}
