+++ src/components/coachsignup/coachstyle.css
@@ -1,7 +1,6 @@
-.gender
-{
+.gender {
     width: 200px;
-    padding:10px 20px;
+    padding: 10px 20px;
     outline: none;
     font-weight: 400;
     border: 1px solid #607d8b;
@@ -12,82 +11,89 @@
     border-radius: 30px;
     margin-bottom: 10px;
 }
-.registerError
-{
+
+.registerError {
     color: red;
 }
-*{
+
+* {
     --main-color: rgb(129, 207, 223);
 }
+
+
 /*#region Login*/
 
+
 /* begin login component */
-section{
+
+.coachSignupContainer {
     position: relative;
     width: 100%;
-    height: 100vh;
+    min-height: 100vh;
     display: flex;
-
-
 }
-section .imgbg {
+
+.coachSignupContainer .imgbg {
     position: relative;
     width: 50%;
     height: 100%;
 }
-section .imgbg::before{
+
+.coachSignupContainer .imgbg::before {
     content: '';
     position: absolute;
     top: 0;
     left: 0;
     width: 100%;
     height: 100%;
-    background: linear-gradient(225deg,var(--main-color),#0f4f94);
+    background: linear-gradient(225deg, var(--main-color), #0f4f94);
     z-index: 1;
     mix-blend-mode: screen;
-
 }
-section .imgbg img{
+
+.coachSignupContainer .imgbg img {
     position: absolute;
     top: 0;
     left: 0;
     width: 100%;
     height: 100%;
     object-fit: cover;
-    
 }
-section .cont{
 
+.coachSignupContainer .cont {
     display: flex;
     justify-content: center;
     align-items: center;
     width: 50%;
-    height: 100%;
-     /* background-color: white; */
+    /* height: 100%; */
+    background-color: red;
 }
-section .cont .formB{
+
+.coachSignupContainer .cont .formB {
     width: 50%;
     height: 100%;
-
+    /* margin to avoid navbar */
+    margin-top: 100px;
+    background-color: blue;
 }
-section .cont .formB h2{
-    color:#607d8b ;
+
+.coachSignupContainer .cont .formB h2 {
+    color: #607d8b;
     font-weight: 600;
     font-size: 1.5em;
     text-transform: uppercase;
-    margin-bottom: 20px;
+    /* margin-bottom: 20px; */
     border-bottom: 4px solid #03a9f4;
     display: inline-block;
     letter-spacing: 1px;
     margin-left: 5px;
 }
 
-section .cont .formB .inputB
-{
-    margin-bottom: 20px;
+.coachSignupContainer .cont .formB .inputB {
+    /* margin-bottom: 20px; */
 }
-section .cont .formB .inputB span
-{
+
+.coachSignupContainer .cont .formB .inputB span {
     font-size: 20px;
     margin-bottom: 5px;
     display: inline-block;
@@ -96,10 +102,10 @@ section .cont .formB .inputB span
     font-size: 16px;
     letter-spacing: 1px;
 }
-section .cont .formB .inputB input 
-{
+
+.coachSignupContainer .cont .formB .inputB input {
     width: 100%;
-    padding:10px 20px;
+    padding: 10px 20px;
     outline: none;
     font-weight: 400;
     border: 1px solid #607d8b;
@@ -108,61 +114,59 @@ section .cont .formB .inputB input
     color: #607d8b;
     background: transparent;
     border-radius: 30px;
-    margin-bottom:10px ;
+    /* margin-bottom: 10px; */
 }
 
-section .cont .formB .inputB input[type="button"]
-{
-background: #03a9f4;
-color: #fff;
-outline: none;
-border: none;
-font-weight: 500;
-cursor: pointer;
+.coachSignupContainer .cont .formB .inputB input[type="button"] {
+    background: #03a9f4;
+    color: #fff;
+    outline: none;
+    border: none;
+    font-weight: 500;
+    cursor: pointer;
 }
 
-
-section .cont .formB .inputB input[type="button"]:hover
-{
-    background: #0088c7;  
+.coachSignupContainer .cont .formB .inputB input[type="button"]:hover {
+    background: #0088c7;
 }
-section .cont .formB .inputB .remember
-{
+
+.coachSignupContainer .cont .formB .inputB .remember {
     margin: 10px;
     color: #607d8b;
     font-weight: 400;
     font-size: 14px;
 }
-section .cont .formB .inputB p
-{
+
+.coachSignupContainer .cont .formB .inputB p {
     color: #607d8b;
 }
-section .cont .formB .inputB p *
-{
-    color:var(--main-color);
 
+.coachSignupContainer .cont .formB .inputB p * {
+    color: var(--main-color);
 }
 
-section .cont .formB h3{
+
+/* .coachSignupContainer .cont .formB h3 {
     color: #607d8b;
     text-align: center;
     margin: 80px 0 10px;
     font-weight: 500;
-}
-section .cont .formB .somed
-{
+} */
+
+
+/* .coachSignupContainer .cont .formB .somed {
     display: flex;
     justify-content: center;
     align-items: center;
-}
-section .cont .formB  .somed  li:hover
-{
+} */
+
+
+/* .coachSignupContainer .cont .formB .somed li:hover {
     background: var(--main-color);
-    
     transform: scale(1.5);
 }
-section .cont .formB .somed li
-{
+
+.coachSignupContainer .cont .formB .somed li {
     list-style: none;
     width: 50px;
     height: 50px;
@@ -171,63 +175,62 @@ section .cont .formB .somed li
     align-items: center;
     background: #e9eced;
     border-radius: 50%;
-    margin: 0 7px ;
+    margin: 0 7px;
     cursor: pointer;
 }
-section .cont .formB .somed li img
-{
-   transform: scale(0.9);
-   
 
-}
+.coachSignupContainer .cont .formB .somed li img {
+    transform: scale(0.9);
+} */
 
-@media(max-width:768px)
-{
-   section .imgbg{
-       position: absolute;
-       top:0;
-       left: 0;
-       width: 100%;
-       height: 100%;
-   }
-   section .imgbg img {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-   section .cont{
-       display: flex;
-       justify-content: center;
-       align-items: center;
-     top:0;
-     left: 0;
-     z-index: 1;
-
-     width: 100%;
-     height: 100%;
+@media(max-width:768px) {
+    .coachSignupContainer .imgbg {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
     }
-    section .cont .formB{
-     width: 100%;
-     padding: 40px;
-     background-color:rgba(255,255,255, 0.9);
-     margin: 50px; 
-     box-shadow: 0px 0px 10px 4px rgba(80, 80, 80, 0.43) !important;
-     border-radius: 20px;  
+    .coachSignupContainer .imgbg img {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
     }
-    section .cont .formB h3 {
+    .coachSignupContainer .cont {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        top: 0;
+        left: 0;
+        z-index: 1;
+        width: 100%;
+        height: 100%;
+    }
+    .coachSignupContainer .cont .formB {
+        width: 100%;
+        padding: 40px;
+        background-color: rgba(255, 255, 255, 0.9);
+        margin: 50px;
+        box-shadow: 0px 0px 10px 4px rgba(80, 80, 80, 0.43) !important;
+        border-radius: 20px;
+        /* margin to avoid navbar */
+        margin-top: 170px;
+    }
+    .coachSignupContainer .cont .formB h3 {
         color: #607d8b;
         text-align: center;
         margin: 20 px0 10px;
         font-weight: 500;
     }
-
 }
-section .overlay{
+
+
+/* .coachSignupContainer .overlay {
     width: 100%;
     height: 100%;
-    background: rgba(255,255,255, 0.8);
+    background: rgba(255, 255, 255, 0.8);
     z-index: -100;
-}
+} */
